pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Personal portfolio website built with Streamlit"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.40.0",
    "Pillow>=10.0.0",
    "fpdf2>=2.8.5",
]

[dependency-groups]
dev = []

